создает начальные модули
